+++ plugins/woocommerce-admin/packages/admin-e2e-tests/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Add Customers to analytics pages tested #7573
 -   Add `waitForTimeout` utility function #7572
+-   Update analytics overview tests to allow re-running the tests.
 
 # 0.1.1